Add UpdateCommentAsync method to CommentController
This commit introduces the `UpdateCommentAsync` method in the `CommentController` class. This new HTTP PUT endpoint allows for updating comments using their ID and a `CommentUpdateDto`. The method processes the response from the `_commentService.UpdateCommentAsync`, returning a `BadRequest` for unsuccessful updates and an `Ok` response for successful ones. Additionally, comments have been added to clarify the code section related to this method.